requirements/dev.txt
Django==3.1.7
djangorestframework==3.12.2
factory-boy==3.2.0
